CNN/requirements.txt
torch==2.2.1+cu118
transformers==4.38.2
Pillow==9.5.0
matplotlib==3.8.0
numpy==1.24.3
pandas==2.1.0
tqdm==4.66.1
datasets==2.18.0
evaluate==0.4.2